COMP: Fix -Wunused-parameter, -Wunused-variable and -Wsign-compare warnings
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=28096
git-svn-id: http://svn.slicer.org/Slicer4/trunk@28096 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/Core/vtkMRMLScene.cxx
@@ -3655,7 +3655,7 @@ void vtkMRMLScene::SetMaximumNumberOfSavedUndoStates(int stackSize)
 void vtkMRMLScene::TrimUndoStack()
 {
   std::list<vtkSmartPointer<vtkCollection> > removedStacks;
-  while(this->UndoStack.size() > this->MaximumNumberOfSavedUndoStates)
+  while(static_cast<int>(this->UndoStack.size()) > this->MaximumNumberOfSavedUndoStates)
     {
     removedStacks.push_back(this->UndoStack.front());
     this->UndoStack.pop_front();
